webpack + fetch
