+++ develop/templates/recordIndex.pug
@@ -15,5 +15,5 @@
 //        script(src="configs/signup-fields.js")
 //        script(src="configs/login-fields.js")
 //        script(src="modules/http.js")
-//        script(src="services/user-servise.js")
+//        script(src="services/user-service.js")
 //        script(src="main.js")
