+++ documentation/files.html
@@ -291,7 +291,7 @@ ul.inheritsList
 			<hr />
 			
 			<div>
-				<h2><a href="symbols/src/public_services_user-servise.js.html">./public/services/user-servise.js</a></h2>
+				<h2><a href="symbols/src/public_services_user-servise.js.html">./public/services/user-service.js</a></h2>
 				
 				<dl>
