+++ package.json
@@ -7,15 +7,25 @@
     "npm": "5.3.0"
   },
   "presets": [
-    "env"
+    "es2015"
   ],
   "dependencies": {
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
     "body-parser": "^1.18.1",
     "cookie-parser": "^1.4.3",
+    "css-loader": "^0.28.7",
     "express": "^4.16.2",
+    "extract-text-webpack-plugin": "^3.0.1",
+    "file-loader": "^1.1.5",
     "morgan": "^1.9.0",
     "pug": "^2.0.0-rc.4",
-    "uuid": "^3.1.0"
+    "style-loader": "^0.19.0",
+    "url-loader": "^0.6.2",
+    "uuid": "^3.1.0",
+    "webpack": "^3.7.1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
@@ -29,6 +39,7 @@
   },
   "scripts": {
     "compile-pug": "pug --client --no-debug ./public/templates/recordIndex.pug",
+    "build": "babel public -d public_ES",
     "start": "nodemon server/index.js && pug --client --no-debug ./public/templates/recordIndex.pug",
     "lint": "eslint ./**/*.js"
   }
