+++ public/bundle.css
@@ -1,0 +1,322 @@
+body {
+    background: url(./background-light-green-30bdf19179a0d73e1e15d962faaa855c.jpg);
+    background-color: #99bc54;
+    width: 98%;
+    height: 95%;
+    min-height: 600px;
+    position: absolute;
+    top: 0;
+    left: 0;
+    overflow: auto;
+}
+/*Общий стиль кнопок, оформление задней панели*/
+
+
+/*div.panel {*/
+    /*background: url('../image/hello-panel.jpg');*/
+
+    /*position: relative;*/
+    /*width: 50%;*/
+    /*min-width: 300px;*/
+    /*margin: 1.5% auto;*/
+    /*border-radius:5%;*/
+    /*border: ridge black;*/
+    /*padding-top: 1%;*/
+    /*padding-bottom: 2%;*/
+/*}*/
+.panel button, input[type=submit] {
+    font-family: cursive;
+    font-size: 250%;
+    text-shadow: darkgray  2px 2px;
+    box-shadow: 10px 5px darkgrey , 12px 7px black;
+    border-radius: 30%;
+    width: 86%;
+    height: 125px;
+    margin-left: 6%;
+    margin-top: 2%;
+    margin-bottom: 2%;
+    border: ridge black;
+}
+.panel button:hover, input[type=submit]:hover {
+    background-color: cadetblue;
+}
+.panel button:active, input[type=submit]:active {
+    margin-left: 8%;
+    box-shadow: none;
+}
+.panel button:focus, input[type=submit]:focus {
+    outline: none;
+}
+
+
+/*Кнопки навигации*/
+button.navigation {
+    position: fixed;
+    margin-top: 2%;
+    margin-left: 3%;
+    height: 10%;
+    width: 15%;
+    min-height: 60px;
+}
+button.navigation:active {
+    margin-left: 5%;
+    margin-top: 3%;
+}
+
+
+/*Профиль*/
+table.profile {
+    font-family: cursive;
+    font-size: 23pt;
+    position: fixed;
+    max-width: 300px;
+    left: 80%;
+    top: 5%;
+}
+
+/*Логотип башен*/
+nav.panel img {
+    position: fixed;
+    height: 80%;
+    bottom: -4%;
+}
+img.left_tower { left: 0; }
+img.right_tower { right: 0; }
+
+
+/*Оформление передней панели*/
+.panel table {
+    min-height: 400px;
+    max-height: 80%;
+    min-width: 250px;
+    font-family: cursive;
+    font-size: 165%;
+    text-shadow: darkgray  2px 2px;
+    width: 80%;
+    height: 80%;
+    border: ridge black;
+    border-radius: 10px;
+    background-color: #b5ce79;
+    margin: 7% auto;
+    padding: 5%;
+}
+.head_table {
+    padding-bottom: 10px;
+    text-align: center;
+    font-size: 30px;
+}
+.div_rules {
+    min-height: 400px;
+    max-height: 80%;
+    min-width: 250px;
+    font-family: cursive;
+    font-size: 165%;
+    text-shadow: darkgray  2px 2px;
+    width: 80%;
+    height: 80%;
+    border: ridge black;
+    border-radius: 10px;
+    background-color: #b5ce79;
+    margin: 7% auto;
+    padding: 2% 5% 5% 5%;
+}
+
+
+/*Формы регистрации и входа*/
+table.input_form {
+    width: 85%;
+    margin: 10% auto;
+}
+.input_form input {
+    width: 80%;
+    border-radius: 25px;
+    padding: 15px;
+    font-size: 16pt;
+}
+.input_form input[type=submit] {
+    padding: 0;
+    margin-top: 6%;
+    font-family: cursive;
+    font-style: normal;
+    font-size: 120%;
+    height: 60%;
+    min-height: 60px;
+}
+.input_form input:focus {
+    outline: none;
+}
+.input_form td {
+    padding: 5px;
+}
+.input_form tr {
+    height: 50%;
+}
+.input_form .alert_message {
+    padding: 10px;
+    text-align: center;
+    box-shadow: -2px -2px red, 2px 2px red, -2px 2px red, 2px -2px red;
+    text-shadow: 1px 1px black;
+    background-color: #ffd1d1;
+    color: red;
+    border-radius: 15px;
+}
+
+
+/*Варнинги*/
+.warning {
+    position: fixed;
+    min-width: 15%;
+    max-width: 25%;
+    min-height: 7%;
+    padding: 2% 2% 3% 2%;
+    left: 2%;
+    top: 4%;
+    text-align: center;
+    font-family: cursive;
+    font-size: x-large;
+    text-shadow: 1px 1px grey;
+    background-color: rgba(255, 0, 0, 0.95);
+    border-radius: 35px;
+    border-color: black;
+    border-style: dashed;
+    border-width: 5px;
+    z-index: 1;
+}
+.warning p {
+    margin-top: -20px;
+    padding: 0;
+    position: relative;
+    float: right;
+    top: 0;
+    left: 0;
+}
+.warning p:hover {
+    text-shadow: 2px 2px grey;
+    cursor: pointer;
+}
+.warning p:active {
+    top: 2px;
+    left: 2px;
+    text-shadow: 0 0 grey;
+}/*бэкграунд для секций меню*/
+.application-header {
+    font-family: fantasy;
+    font-size: 100px;
+    position: relative;
+    width: 50%;
+    margin-left: 27%;
+    min-width: 300px;
+}
+
+section.section {
+    /*background: url('../image/hello-panel.jpg');*/
+    position: relative;
+    width: 50%;
+    min-width: 300px;
+    margin: 1.5% auto;
+    border-radius:5%;
+    border: ridge black;
+    padding-top: 1%;
+    padding-bottom: 2%;
+}
+
+.section_title {
+    font-family: fantasy;
+    position: relative;
+    margin-left: 20%;
+    margin-right: 20%;
+    width: 60%;
+    align-self: center;
+}
+
+.section button, input[type=submit] {
+    font-family: cursive;
+    font-size: 250%;
+    text-shadow: darkgray  2px 2px;
+    box-shadow: 10px 5px darkgrey , 12px 7px black;
+    border-radius: 30%;
+    width: 86%;
+    height: 125px;
+    margin-left: 6%;
+    margin-top: 2%;
+    margin-bottom: 2%;
+    border: ridge black;
+}
+
+/*Профиль*/
+div.profile {
+    font-family: cursive;
+    font-size: 23pt;
+    position: fixed;
+    max-width: 300px;
+    left: 80%;
+    top: 5%;
+}
+
+
+input {
+    display: block;
+    width: 60%;
+    margin: 0 auto;
+    font-size: 16px;
+    margin-bottom: 3px;
+}.message {
+    font-family: fantasy;
+    position: relative;
+    margin-left: 20%;
+    margin-right: 20%;
+    width: 60%;
+    align-self: center;
+}input.login_input {
+    width: 80%;
+    border-radius: 25px;
+    margin-left: 10%;
+    margin-right: 10%;
+    padding: 15px;
+    font-size: 16pt;
+}
+
+input.login_input [type=submit] {
+    padding: 0;
+    margin-top: 6%;
+    font-family: cursive;
+    font-style: normal;
+    font-size: 120%;
+    height: 60%;
+    min-height: 60px;
+}
+
+input.login_input :focus {
+    outline: none;
+}
+  table.input_form {
+    width: 85%;
+    margin: 10% auto;
+}
+input.sign_up_input {
+    width: 80%;
+    border-radius: 25px;
+    margin-left: 10%;
+    margin-right: 10%;
+    padding: 15px;
+    font-size: 16pt;
+}
+input.sign_up_input [type=submit] {
+    padding: 0;
+    margin-top: 6%;
+    font-family: cursive;
+    font-style: normal;
+    font-size: 120%;
+    height: 60%;
+    min-height: 60px;
+}
+input.sign_up_input :focus {
+    outline: none;
+}
+input.sign_up_input td {
+    padding: 5px;
+}
+input.sign_up_input tr {
+    height: 50%;
+}
+/*# sourceMappingURL=bundle.css.map*/
