+++ public/index.html
@@ -4,29 +4,11 @@
 <head>
     <meta charset="UTF-8">
     <title>Tower Defense</title>
-    <link rel="stylesheet" type="text/css" href="CSS/index.css">
-    <link rel="stylesheet" type="text/css" href="CSS/new.css">
-    <link rel="stylesheet" type="text/css" href="blocks/form/index.css">
-    <link rel="stylesheet" type="text/css" href="blocks/message/index.css">
-    <link rel="stylesheet" type="text/css" href="configs/signup-fields.css">
-    <link rel="stylesheet" type="text/css" href="configs/login-fields.css">
+    <link rel="stylesheet" type="text/css" href="./bundle.css">
 </head>
 
 <body>
-<!--<div id="application"></div>-->
-<script src="blocks/block/index.js"></script>
-<script src="blocks/form/index.js"></script>
-<script src="blocks/message/index.js"></script>
-<script src="templates/scoreBoard.js"></script>
-<script src="blocks/scoreboard/index.js"></script>
-<!--<script src="blocks/scoreboard/scoreboard.pug"></script>-->
-<!--<script src="blocks/scoreboard/compile.js"></script>-->
-
-<script src="configs/signup-fields.js"></script>
-<script src="configs/login-fields.js"></script>
-<script src="modules/http.js"></script>
-<script src="services/user-servise.js"></script>
-<script src="main.js"></script>
+<script src="./out.webpack.js"></script>
 </body>
 
 </html>
